Mandatory part
Falta corregir errores de ejecucion y compilacion

+++ Libft/ft_strlen.c
@@ -6,13 +6,12 @@
 /*   By: kroyo-di <kroyo-di@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 10:31:09 by kroyo-di          #+#    #+#             */
-/*   Updated: 2024/06/18 12:15:12 by kroyo-di         ###   ########.fr       */
+/*   Updated: 2024/07/03 21:28:03 by kroyo-di         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
